style: move contexts to same file
